+++ packages/components/src/color-palette/styles.ts
@@ -12,7 +12,7 @@ import { CONFIG } from '../utils';
 export const ColorHeading = styled( Heading )`
 	text-transform: uppercase;
 	line-height: 24px;
-	font-weight: ${ CONFIG.fontWeightSemibold };
+	font-weight: ${ CONFIG.fontWeightEmphasis };
 	&&& {
 		font-size: 11px;
 		margin-bottom: 0;