+++ routes/template-list/add-new-template/index.tsx
@@ -136,7 +136,7 @@ function TemplateListItem( {
 				>
 					<WCText
 						align="center"
-						weight="var(--wpds-typography-font-weight-semibold)"
+						weight="var(--wpds-typography-font-weight-emphasis)"
 						lineHeight={ 1.53846153846 } // 20px
 					>
 						{ title }