+++ tools/stylelint/config.js
@@ -26,14 +26,23 @@ module.exports = {
 			{
 				'/.*/': [
 					'/--wp-components-color-/',
+					'/\\$font-weight-regular/',
 					'/\\$font-weight-medium/',
 				],
 				cursor: [ 'pointer' ],
 			},
 			{
 				message: ( property, value ) => {
-					if ( value.includes( '$font-weight-medium' ) ) {
-						return '`$font-weight-medium` is kept for backwards compatibility only. Use `$font-weight-regular` or `$font-weight-semibold` based on the intended emphasis.';
+					if (
+						value.includes( '$font-weight-regular' ) ||
+						value.includes( '$font-weight-medium' )
+					) {
+						const variable = value.includes(
+							'$font-weight-regular'
+						)
+							? '$font-weight-regular'
+							: '$font-weight-medium';
+						return `\`${ variable }\` is kept for backwards compatibility only. Use \`var(--wpds-typography-font-weight-default)\` or \`var(--wpds-typography-font-weight-emphasis)\` based on the intended emphasis.`;
 					}
 					if ( property === 'cursor' ) {
 						return 'Use the `var( --wpds-cursor-control )` token for interactive non-link controls. If this is for a link, you can disable this rule.';